feature(#681): move autoconfiguration to starter module
