+++ digiwf-libs/digiwf-spring-security/digiwf-spring-security-core/pom.xml
@@ -16,6 +16,12 @@
             <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
 
+        <!-- feign client -->
+        <dependency>
+            <groupId>io.github.openfeign</groupId>
+            <artifactId>feign-httpclient</artifactId>
+        </dependency>
+
         <!-- Spring security -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
