+++ digiwf-process/digiwf-process-api/src/main/java/de/muenchen/oss/digiwf/process/api/config/impl/ProcessConfigClient.java
@@ -1,18 +1,28 @@
 package de.muenchen.oss.digiwf.process.api.config.impl;
 
 import de.muenchen.oss.digiwf.process.api.config.api.dto.ProcessConfigTO;
+import de.muenchen.oss.digiwf.spring.security.client.DigiwfFeignOauthClientConfig;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
+/**
+ * ProcessConfigClient is the feign client to obtain the process configuration.
+ */
 @FeignClient(
         name = "${feign.client.config.digiwf-process-config.name:digiwf-process-api}",
         url = "${feign.client.config.digiwf-process-config.url:${feign.client.config.default.url:http://localhost:39146}}",
-        configuration = FeignConfig.class
+        configuration = DigiwfFeignOauthClientConfig.class
 )
 public interface ProcessConfigClient {
 
+    /**
+     * Feign Client to obtain the process configuration from digiwf-engine.
+     *
+     * @param engine
+     * @return
+     */
     @RequestMapping(method = RequestMethod.GET, value = "/rest/processconfig/{key}", consumes = "application/json")
     ProcessConfigTO getProcessConfig(@PathVariable("key") final String engine);
